+++ src/runtime/components/icons/chains/MIconChainRise.vue
@@ -7,13 +7,13 @@ defineProps<{
 
 <template>
   <img
-    src="https://green-obedient-albatross-155.mypinata.cloud/ipfs/bafkreiam5jndlyhh4owaktg56wofifdmm5x5v7kdvsndgjzoidzyzocskq"
+    src="https://green-obedient-albatross-155.mypinata.cloud/ipfs/bafkreiatvb4wduh2ek5kdszcro264xbsmkk67pneksrry4b2jissljz3im"
     :width="size ?? 24"
     :height="size ?? 24"
-    :alt="alt || 'Soneium'"
-    class="rounded-full"
+    :alt="alt || 'Rise'"
     loading="lazy"
     decoding="async"
     referrerpolicy="no-referrer"
+    class="rounded-full"
   />
 </template>